backend/requirements/base.txt
django>=4.2.16
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
channels>=4.1.0
daphne>=4.1.2
celery>=5.4.0
redis>=5.0.8
channels-redis>=4.2.0
django-cors-headers>=4.4.0
django-filter>=24.3
python-decouple>=3.8
openai>=1.51.0
requests>=2.32.3
drf-spectacular>=0.27.2
Pillow>=10.4.0
